requirements.txt
altair==5.0.1
amazon-textract-caller==0.0.29
amazon-textract-prettyprinter==0.1.3
amazon-textract-response-parser==1.0.0
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
boto3==1.28.25
botocore==1.31.25
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
decorator==5.1.1
executing==1.2.0
fastapi==0.101.0
gitdb==4.0.10
GitPython==3.1.32
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
numpy==1.25.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.39
protobuf==4.24.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pydantic==2.1.1
pydantic_core==2.4.0
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
s3transfer==0.6.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.27.0
streamlit==1.25.0
tabulate==0.9.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
uvicorn==0.23.2
validators==0.21.2
watchdog==3.0.0
wcwidth==0.2.6
zipp==3.16.2
